Time: 1973 ms (10.55%), Space: 370.2 MB (5.83%) - LeetHub

+++ 0446-arithmetic-slices-ii-subsequence/0446-arithmetic-slices-ii-subsequence.cpp
@@ -3,7 +3,7 @@ public:
     int numberOfArithmeticSlices(vector<int>& nums) {
         int n = size(nums);
         long long ans = 0;
-        map<pair<int, long long>, long long> dp;
+        map<pair<int, long long>, int> dp;
         for(int i = 0; i < n; i++) {
             for(int j = 0; j < i; j++) {
                 long long diff = 1LL * nums[i] - nums[j];
